MB-5723 Allow log level to be configurable
**Description**

It is a STIG requirement for the application log levels to
be configurable, and for them to default to `INFO`. It is also an
industry best practice. Rails, for example, provides this out of the
box.

**How**

* Add new logging level flags to `pkg/cli/logging.go`

* Modify our existing `logging/log.go` package to accept a string that
defines the desired log level, based on the `LOGGING_LEVEL` env var.
Previously, we were limiting the configuration to the default log
levels provided by zap (`DEBUG` for development, and `INFO` for
production), and the ability to set the log level to `DEBUG` via a
boolean based on the `DEBUG_LOGGING` env var defined in
`pkg/cli/verbose.go`.

* Update all calls to `logging.Config` to replace the boolean with
the string from the new logging level flag.

* Remove `pkg/cli/verbose.go` and associated tests.

* Remove `DEBUG_LOGGING=true` from prd, stg, and exp `.env` files, and
replace with `LOGGING_LEVEL=info`, which is the desired default level
in deployed environments per the STIG.

* Make the logging levels case-insensitive because it's the nice
thing to do

This allows us to change the log level to any of the 5 acceptable
values (`FATAL`, `ERROR`, `WARN`, `INFO`, `DEBUG`), without having to
deploy new code. We only need to change the env var and restart the
app.

**Reviewer Notes**

1. From the master branch, run `make server_run` locally
- [ ] Verify that you see `DEBUG` entries in the log, such as
"Server DOD Key Pair Loaded"
3. Stop the server, and switch to this branch
4. Run `make server_run`
- [ ] Verify that you only see log items marked as `INFO`.
6. Add `export LOGGING_LEVEL=DeBUG` to your `.envrc.local`
7. Stop the server
8. Run `direnv allow`
9. Run `make server_run`
- [ ] Verify that you see `DEBUG` entries
10. Update the log level in your `.envrc.local` to a bogus value
11. Repeat steps 5-7
- [ ] Verify that the server crashes and lets you know that the
value is not allowed, and what the allowed values are.