Don't hardcode LOGGING_LEVEL in Makefile
This was an inadvertent change in the spirit of preserving previous
functionality, but we no longer want to hardcode the LOGGING_LEVEL
when running the server. We want folks to be able to override it in
their `.envrc.local`. Otherwise, this prevents us from testing the
`LOGGING_LEVEL` functionality locally.