MB-5722 Make stack trace length configurable
One of the STIG requirements is to limit the number of lines in a
stack trace to only show the first 3 code paths, and to make the
number of lines configurable, in case we need to increase or decrease
it. The configuration should be via an ENV var that can be changed
without having to deploy new code. This requires creating a new AWS
SSM parameter for this, which the Infra team kindly already took care
of.

**How**

* In order to get access to the stack trace, we need to define a new
custom Zap encoder, and do the stack trace manipulation in the
encoder's `EncodeEntry` function. We need to create two custom
encoders because development uses the console encoder, and deployed
environments use the JSON encoder.

* Instead of adding yet another parameter to the logging `Config`
function, which is a code smell, especially when there are multiple
parameters of the same type, I refactored `Config` to accept
functional options instead. Advantages:
- The order of the options doesn't matter, so no more risk of messing
things up
- You don't have to pass in all the options if you don't need them
all. This means if we add a new option in the future that is only
needed in `serve.go`, we won't have to make changes everywhere
`logging.Config` is called
- Options with descriptive names makes it clear what you are passing

* In order to limit the stack trace to the first 3 code paths, and
also keep it useful for troubleshooting, I modified the stack trace
so that it only shows code paths within the "mymove" packages.

**Bonus**
I updated the dev logger to colorize the log levels, so it's easier
to spot the ERRORs in red.

**Reviewer Notes**

1. Replace `pkg/services/move_order/move_order_fetcher.go` with the contents of [this gist](https://gist.github.com/monfresh/c04364b04c47387c95e0dd4f1ed90fb9)
1. Run `make server_run`
2. In a new terminal tab, run `make office_client_run`
3. Use the local sign in to sign in with the TOO user
- [ ] Verify that you see a stack trace, and that it only contains
code paths from "mymove/pkg", and that it only shows 6 lines by
default
4. Add `export STACKTRACE_LENGTH=10` to your `.envrc.local`
5. Stop the server
6. Run `direnv allow`
7. Run `make server_run`
8. Bring focus back on the browser tab/window that has the office app
- [ ] Verify that the stack trace is now 10 lines long
9. Update STACKTRACE_LENGTH in your `.envrc.local` to a string instead of an integer
10. Repeat steps 5-8
- [ ] Verify that the full stack trace is shown (but still only with mymove code paths)

Ideally, if we pass in something that's not an integer, it should blow up.
I'm trying to figure out why it's not.