Use structs for invoice Header and ServiceItems
The service items have the same segments repeated for each item, followed
by a single L3 segment.
I made a struct for the repeated segments, and moved the L3 segment out
to the Invoice858C struct.

We also decided not to use the linkedhashmap, and instead just make a struct
for all the header items. It's pretty long and some are optional, which
is a bit boilerplatey, but so far I think it's the right approach.