MB-5696 Fix bugs around updating a Move's status
**Description**
Service items can only be created on a Move whose status is either
Approved or Approvals Requested. However, this was not properly
enforced in `mto_service_item_creator`. The status update on the move
was performed outside of the transaction that created the service
items. So if service items were created on an invalid move, even
though the status update failed, the service items were still created.

I also found a bug in `mto_service_item_updater`, where if a service
item's status was set back to SUBMITTED, the move's status wouldn't
get reset to APPROVALS REQUESTED, and so the TOO wouldn't know to
review it.

To fix the first bug, I added a check at the beginning of the
service item creation function to check and fail early if the move's
status is not 'Approved' or 'Approvals Requested', because if it's
not, nothing in the rest of the function matters. I also updated the
error message to be more friendly for the Prime.

I added several missing tests around these two bugs, and I plan on
adding more, as well as doing some refactoring to make the code
easier to understand.