MB-6069 Don't display contents of error parameter
In an attempt to improve the user experience when something goes wrong
with the login.gov authentication, instead of returning a 500 error,
the user was sent back to the sign in page with an error message:
"There was an error during your last sign in attempt. Please try again."

For some reason, the original implementation also passed along an
error code via the "error" URL query parameter, and displayed that
error code to the user. There were only two possible error codes:
"INVALID_REQUEST" and "UNKNOWN_ERROR". I don't see how this additional
information helps the user. It seems more confusing than helpful.

Since the frontend simply displays the contents of the error message,
anyone could devise a malicious error message and fool someone into
clicking a milmove URL with the error parameter, which is the content
injection vulnerability we are fixing here.

The easiest way to fix this is to not show the contents of the error
parameter, since it doesn't seem to add any value.

Here is the original PR that introduced the vulnerability:
https://github.com/transcom/mymove/pull/1738