Add Danger rule for suite.DB().TruncateAll()
As part of Infra's work to ensure our local DB permissions match those
in our deployed environments, we changed the truncation call from
`suite.DB().TruncateAll()` to `suite.TruncateAll()`. As a result,
`suite.DB().TruncateAll()` no longer works, and doesn't throw an
error, which can lead to confusion and broken tests.

This Danger rules serves as a reminder of the change.