Only set ServiceMemberID in session if app is mil
Recently, we added a header on the Office App to display the Office
User's name and GBLOC. When this was first implemented, the
information was only displaying for some users, but not others. This
information comes from the payload of the `/users/logged_in` endpoint.

The reason why the header wasn't working for some people, namely those
who existed as both service members and office users, is because the
`ShowLoggedInUserHandler` was only returning the Office User info if
`!session.IsServiceMember()`. Unfortunately, this only worked if the
current user had never signed into the service member application.
This is because when an office user signs in (and if they also have a
service member account), we populated the session with their Service
Member ID, which makes `session.IsServiceMember()` return `true`.

So, one reasonable solution, and the simplest one, is to only set the
ServiceMember ID in the session if the current app is the `mil` app.