MB-6604 Populate MilSessionID on first login
This PR fixes a bug where the `current_mil_session_id` field in the
Users table was not being populated on the very first sign in. That
meant that we couldn't revoke a User's session unless they signed out
and back in.

The reason why the field was not being populated is because this action
depended on `session.IsServiceMember()` being `true`. The Users table
has session fields for all three apps (mil, admin, office), and the
auth code determines which field to populate based on the type of user.

The problem was that `IsServiceMember` is only true if the session
contains the ID of the service member tied to the user, but the
service member was not created until after the authentication was
complete. The auth function only created the user, then redirected to
the onboarding home page, which then created the service member (this
happens in `src/sagas/onboarding.js`).

In this PR, the auth code now creates the service member right after
creating the user, and it populates the session with `ServiceMemberID`,
which allows `session.IsServiceMember()` to be `true`, and populates
the `current_mil_session_id` field.