MB-6437 Add script/data to benchmark doc viewers
We need to ensure that the document viewer is performant & reliable in
the transportation offices, which are known to have bandwidth issues.
This PR adds a script to populate a single move with 3 files of
varying sizes, that both the TOO and TIO can view: orders document
viewer and proof of service document viewer.

To run the script:

```
make db_bandwidth_up
```

To verify the docs were created and can be viewed, follow these steps:

1. Start the server and client:

```
make server_run
make office_client_run
```

2. Click on "Local Sign in" when the office app appears
3. Sign in with the single office user that the script created, and
view the docs as a TOO, and then as a TIO.

As a TOO:

1. Click on the only move in your queue
2. Click on "Orders" in the left sidebar
3. Click on "View & edit orders" in the top right

As a TIO:

1. Click on "Change user role" in the top left
2. Click on "Select transportation_invoicing_officer"
1. Click on the only payment request in your queue
2. Click on "Review service items"