MB-6673 Speed and clean up common Make commands
This PR updates `db_dev_e2e_populate` so that instead of resetting and
migrating the DB from scratch, it runs only the missing migrations, if
any, then truncates the tables. This saves 3 minutes each time you
run it. The vast majority of the time that someone runs this command,
their local dev DB is already set up, so it doesn't make sense to
destroy the DB and rebuild it from scratch when all you need is to
clear the tables and repopulate them.

And in the case where you do need to start from scratch for some reason,
this PR introduces a `db_dev_fresh` command that will reset the DB and
run all the migrations.

To follow this pattern, I updated the existing `e2e_test` script so
that it only clears the test DB tables and repopulates them, which
makes it a whole lot faster to rerun end-to-end tests with fresh data.
And in the case where you do need to start from scratch, I introduced
the `e2e_fresh` script that does what `e2e_test` used to do.

I also moved `db_dev_e2e_populate` to the `DB_DEV TARGETS` section,
where it belongs since it's used to populate the Dev DB, and not for
running automated E2E tests.