Remove unused context function argument
About 2 years ago, we added instrumentation via Honeycomb, which
required passing in the context to functions. Since then, we've removed
Honeycomb, but we didn't remove the context argument. Here are example
PRs that introduced the context argument:

https://github.com/transcom/mymove/pull/1269
https://github.com/transcom/mymove/pull/1326

This PR removes all unused code related to the context argument that
was added at the time Honeycomb was introduced.