Add new service item param keys from pricing
We need to save information looked up by pricers with the service item
param keys so that we can look up the exact calculation that was
used at a later date, even if our pricing tables change.

We are doing this to enable future changes that will show the TIO
how we calculated each price.