MB-6935 Set access code flag in auth context
When the service member is first created upon their very first sign in,
we set their `requires_access_code` field to match the value of the
`FEATURE_FLAG_ACCESS_CODE` env var. This env var was not being properly
read in `auth.go`, and so it was always false, which meant customers
in prod could bypass the access code requirement.

The fix was to set the env var in the `context` that `auth.go` uses,
which is different from the `HandlerContext` that handlers use.