MB-6901 Approve Move after TOO approves Shipment
When a TOO approves a shipment, there are default service items that get
created, and they are approved by default. When a move doesn't have any
service items in SUBMITTED status, the move's status should be
`Approved`, but in `mto_service_item_creator`, we were always setting
the move's status to `Approvals Requested`.

Instead, we should be iterating through all of the move's service items
to check their status, and only if at least one service item is in
`Submitted` status should we set the Move's status to
`Approvals Requested`. Otherwise, we should set it to `Approved`.