MB-6917 Add PaymentServiceItemParams to GHC API (#6020)
* MB-6917 Add PaymentServiceItemParams to GHC API

A TIO needs to see how weights factor into the pricing of service items
so that they can understand how the price was determined.

This information comes from the `payment_service_item_params` and
`service_item_param_keys` tables.

The GHC YAML file already included the necessary fields in the
definition of the `getPaymentRequestsForMove` endpoint via the
definitions for
PaymentRequest->PaymentServiceItem->PaymentServiceItemParam

To add these to the payload, I copied the functions from the Prime API,
and then I updated the `FetchPaymentRequestListByMove` function to
perform the necessary eager loading to fetch the data from the
DB.

I used TDD to write a failing test first, and converted the existing
`TestGetPaymentRequestsForMoveHandler` test to an integration test that
hits the DB, as opposed to using mocks. To figure out how to properly
set up the DB with all the necessary data, I looked up existing usage of
`paymentServiceItemParams` in our codebase, and found one in the Prime
API's `pkg/handlers/primeapi/move_task_order_test.go`. I copied that
over and added tests for the `Key` and `Value` fields in the payload.

It's not clear from the JIRA issue what the payload needs to contain for
the client to display the information to the TIO. If it needs the full
paylod as defined in the `PaymentServiceItemParam` section of the YAML,
then I'll add more tests for the rest of the payload.