MB-6928 Get ServiceItemParams for payment request
**Description**

A TIO needs to see how weights factor into the pricing of service items
so that they can understand how the price was determined.

The payload was already modified in PR #6020 to include service items
params, so in this PR, all we needed to do was eager load the service
items params and key tables.

I used TDD to write a failing test first, and converted the existing
`TestFetchPaymentRequestHandler` test to an integration test that
hits the DB, as opposed to using mocks. I used the same test setup as
in PR #6020.

**Reviewer Notes**

Note that this PR only implements the backend portion of this feature.
There is a separate unscheduled story (in TTV's March Slice) for the
[frontend portion](https://dp3.atlassian.net/browse/MB-4800).