MB-7015 Prevent Move approval if Order is invalid (#6171)
* MB-7015 Validate presence and format of TAC

When an order is first created during onboarding, it's OK for the TAC
to be nil because we don't know it yet. However, once the move is
submitted, a TOO cannot update the move until they fill out the TAC.
In addition, the TAC must be exactly four alphanumeric characters.