Updating hash to be the one from the master branch
This PR is currently using the hash of the original commit on the circleci-docker dependabot branch for the corresponding cypress version update there.  That branch hash exists in docker (the image built at the time the commit was made), but does not exist anymore in the repo since we squash and merge dependabot updates.  Plus, it was a week ago when that docker image was built so we would be potentially missing other changes on the master branch since then if we use that image.  Looks like our previous hash updates have been using master branch hashes, so going to update this to do the same.