Remove readFromSyncada param & use viper for env vars
I had added a param to to the ProcessReviewedPaymentRequests support API endpoint
but on consulting with the team decided to just attempt to read from Syncada on
every run.

I also changed to using the existing cli function to create an ssh connection
because it seemed weird to setup a viper object in an api endpoint, but I
decided to just do it instead of duplicating the env variable reading code.