Fix flakey admin Cypress tests
I believe the flakey tests we have been seeing in `adminUsers.js` and
`webhookSubscriptions.js` are due to them always being together in the
same parallel run. From what I can tell, all builds are running the
exact same groups of tests in the same order and in the same parallel
runs.

It just happens that the only parallel run with 2 tests that are both
from the admin folder is Parallel Run 0, which runs both `adminUsers.js`
and `webhookSubscriptions.js` one after the other. And because we had
backend code that would redirect to the admin home page if a session
already existed, and because Cypress did not automatically clear the
session in between tests, these were the tests that failed often.

When then added calls to clear cookies before the webhook subscription
tests, but the order in which the call to clear the cookies was made
got botched during a merge conflict. In the second webhook subscription
test (Details Show Page), cookies were cleared before preparing the
admin app, which cause this test to start failing again.

To hopefully solve this once and for all, I've removed the backend
redirect because it causes more confusion than it does anything useful.
I also replaced `clearAllCookies` with `cy.logout()` and I added
calls to `logout` in other admin tests for good measure. With the
backend change, it shouldn't be necessary to logout but I added it as
an extra safety measure. The `clearAllCookies` command we added uses an
undocumented and unsupported Cypress option, so I thought it's best to
stick with our custom `logout` command.