Update ISA validation to allow Syncada 997 files
we were reading AuthorizationInformation and SecurityInformation as spaces
and so they were using the eq validations intead of getting skipped by omitempty.
Trimming the fields when they are parsed ensures that they will be seen as empty.

Syncada also uses a colon for the ComponentElementSeparator in 997s, so I added
that as an option.