MB-7112 Add service counseling Move statuses
This PR adds two new Move statuses to support Service Counseling:
- MoveStatusNeedsServiceCounseling
- MoveStatusServiceCounselingCompleted

The Move status column comment was updated to explain how these new
statuses work as part of the end to end workflow.

Note that to make the long status names readable, we made them Pascal
case as opposed to capitalized like the existing ones. If we like this
new convention, we can update the case of the older ones.

For readability, we formatted the column comment with newlines and
paragraphs. It doesn't seem to have any adverse effects in TablePlus,
but in `psql` it seems harder to read. We are looking for feedback
from folks with more experience with comments.