MB-7635 Initial refactoring of MoveRouter
This renames `MoveStatusRouter` to a more generic and meaningful
`MoveRouter`, and renames `RouteMove` to `Submit`, which is what the
`SubmitMoveHandler` now calls, since this happens when the customer
submits their move.

The `Submit` function in the `MoveRouter` then calls internal functions
that determine whether to send the move to the Service Counselor or
TOO. These are internal functions because there aren't any other places
in the app where we would want to directly send a move to a SC or TOO.

I also copied over the other Move state machine functions from the Move
model, but didn't delete them from the model yet, to keep things
working, and to make this initial commit easier to review.