MB-7669 Allow SC Completed Moves to be approved
The conditional in `MakeAvailableToPrime` was only allowing Moves in
`Submitted` status to be approved. This meant that Moves in the new
`ServiceCounselingCompleted` status didn't get a chance to get approved,
but the code still tried to create service items, which failed because
the Move wasn't approved yet.

Instead of wrapping the call to `move.Approve` in a conditional, we
should call it directly since it has its own logic for determining
when a move can be approved, and it will return an error if the move
can't be approved, which will prevent service items from being created.