Refactor PaymentRequestReviewedProcessor and tests
This speeds up the tests by more than half: from over 30 seconds down to
about 12 seconds.

Here are the principles I used in this refactor:

- Functions should be as short as possible
- Extract each logical component of a function into a separate function
- Where it makes sense, create an interface for that logical component
so it can be mocked easily in tests
- Tests should be as fast as possible
- Tests should use the DB as little as possible
- Don't test the same things in every test. Each variation should focus
on what's unique and mock everything else that was already tested.
- If there are more than 2 subtests that write to the same DB table,
use `suite.TruncateAll()` before each subtest to avoid confusion around
what data is expected to be present in each subtest. By not truncating,
it prevents running a single subtest and forces the developer to run
the entire set of subtests, which slows them down. This can be avoided
by using the DB as little as possible. See the previous principles.

I added more comments inline.

There is further improvement that can speed up the tests even more, by
being able to mock more of this flow.

There's also a missing test that I wasn't sure how to write off the top
of my head. Since this workflow runs while iterating through payment
requests, we want to test that if there is more than one payment
request, if the first one succeeds, but the second one does not, the
first one should pass, but the second one should not. I think that's
what we want, but I'm not 100% sure. I don't know if testify supports
this kind of mocking, where you tell it to return an error only on the
second time.