MB-8077 Speed up models tests
The main purpose of this PR is to speed up the models tests by 4x:
from about 80s down to about 20s. The main reason our server tests are
slow is because we truncate the DB in between each test. In this PR, we
remove the call to suite.TruncateAll from the SetupTest function, and
only call it in tests that require it.

This PR also refactors models tests to remove unnecessary DB calls by
calling `Validate` instead of `ValidateAndCreate` or `ValidateAndSave`,
and by using the stubbed versions of testdatagen objects.

I also added the `gotestsum` tool, which provides a nicer way to run
tests in the console. For example, it can display the output by package
name, with the timings for each package, and without the error logging
that can obscure real test failures. It also provides an easy way to
check the slowest tests based on a threshold. More usage details here:
https://github.com/gotestyourself/gotestsum