MB-7358 Add orders details for Services Counselor
This adds the Orders details definition list to the Services Counseling
view. A Services Counselor needs less info than a TOO, so I duplicated
the `OrdersList` that we use on the TOO's view, and removed the info
that a Services Counselor does not need to see.

I also added the `View and edit orders` button, but it doesn't do
anything currently, as that will be implemented in a separate story.