Fix bug with multiple quotes at end of sql string
Statements with escaped quotes at the end of a string would break splitting:
COMMENT ON COLUMN foo IS 'a ''comment''';
When the splitter got to the semicolon, it would still think it was
within a quoted string.
I added a quick fix to this that I intend to improve in a later commit.