Update unique constraint to allow different edi types with same ICN
The payment_request_to_interchange_control_number table stores an icn
tied to a payment request, but there maybe the same icn but for
different edi types this commit makes it so that can be saved vs
violating the constraint.

This is needed because an edi of a different type may have the same icn
as another. since icn is generated by the sender of the file.