Fix MTOShipment test
This test was panicking because the handler is receiving
`services.InvalidInputError` from the service, but without validation
errors, and then the handler calls `payloadForValidationError`, which in
turn calls `NewValidationErrorsResponse` which tries to iterate through
`range verrs.Keys()`, but `verrs` is nil.

The reason the validation errors are nil is because by the time the
payload has arrived at the `CreateMTOShipment`, the
`shipment.PickupAddress` is `nil`, and so the service is returning

```golang
services.NewInvalidInputError(
  uuid.Nil, nil, nil,
  "PickupAddress is required to create an HHG or NTS type MTO shipment"
)
```
with `nil` validation errors.

It looks like the GHC API is not following the same pattern as other
APIs. Instead of defining a `payloadForValidationError`, the other APIs
define a `ValidationError` and `ClientError` in `model_to_payload`, and
then the handler calls `payloads.ValidationError` and passes in
`e.ValidationErrors`, which works even if the validation errors are nil
because in the `ValidationError` function, it only includes
`InvalidFields` if validation errors are present.

The other difference is that the other APIs expect the `ValidationError`
to contain an array of errors for each invalid field, but in `ghc.yaml`,
we just pass in a string. We should probably make all APIs consistent.

For now, the easiest fix is to add a conditional to
`payloadForValidationError` so that it only adds `InvalidFields` if
validation errors are present. And then the test should expect that
the invalid fields are empty. The reason why the test used to pass
with non-empty invalid fields is probably because it was not doing what
we thought it was testing.