Fix bug that nullifies shipment ApprovedDate
In PR #4686, we introduced a bug that nullified a shipment's
`ApprovedDate` if a `ScheduledPickupDate` was included in the payload.
This affects approving diverted shipments because we determine which
endpoint to hit based on the presence of the `ApprovedDate`.

When a shipment is diverted, the original shipment that was created by
the customer gets updated by the Prime to set its `Diversion` field to
`true`. In addition, the Prime creates a new shipment. Both of these
shipments have the `SUBMITTED` status, but only the new shipment should
have service items automatically created for it.

To distinguish between these two scenarios, we created separate
endpoints in the GHC API: `shipments/{shipmentID}/approve` for new
shipments, and `shipments/{shipmentID}/approve-diversion` for existing
shipments that had previously been approved by the TOO. The way the
front end determines which endpoint to call is by looking at the
shipment's `ApprovedDate` because only the previously-existing shipment
would have that set. The new shipment hasn't been approved yet by the
TOO.

But because of this bug, both shipments end up with a nil
`ApprovedDate`, and the frontend sends both to the `approve` endpoint,
which creates service items, and so the existing shipment ends up with
duplicate service items.