MB-8645 Remove image from crating service items
We originally thought we might need to support uploading images for
crating service items, but that's not the case anymore, so we're
removing references to crating images from the front-end. The backend
portion never got implemented, so there's nothing to remove there.

Note that we still have styles and component names with image
references, like `.detailImage` and `ServiceItemTableHasImg`, so those
should probably be renamed.