MB-8665 Move order validations to order_updater (#6995)
* MB-8665 Move order validations to order_updater

When we first added validations to prevent a TOO from submitting a move
while certain order fields were missing, we made the mistake of putting
the validations in the Order model as opposed to the service. Back then,
we only had one generic service for updating orders, so it might have
made sense at the time.

Since then, we've added more complexity, such as requiring separate
validations for counselors versus TOOs, and allowing customers to
upload amended orders even if the move has made it past service
counseling but hasn't been updated by the TOO yet.

Instead of trying to fit all of these different scenarios into a
generic validation in the Order model, it now makes more sense to
only add the validation where it applies, which is only when a TOO is
updating an order and when they are approving a move.