Fix createHHGWithPaymentServiceItems in devseed
In PR #6995, we updated the service item creator to allow setting its
DB connection. This was necessary to allow the service item creator to
use the same transaction as the one in `MakeAvailableToPrime`.

The problem is that anytime `MakeAvailableToPrime` is called, the
service item creator's connection will be set to the transaction that
was created in `MakeAvailableToPrime`, and it will stay that way unless
it gets set again.

In `createHHGWithPaymentServiceItems`, we first call
`MakeAvailableToPrime`, and then later on, we approve the shipments by
calling `UpdateMTOShipmentStatus`, which in turn calls the service item
creator, which is now using the transaction connection. However, the
move that's associated with the shipments was created with the initial
DB connection (not a transaction) that was passed into the devseed, and
so it blows up when looking up the move associated with the
shipment.

To fix this, we need to set the service item creator's connection back
to the initial DB connection so that `UpdateMTOShipmentStatus` can work.

This is by no means ideal, and is a result of us not implementing DB
connections The Go Way, and due to our choice (back when this project
started) of not using more established frameworks. We are planning a
refactor to straighten things out.