Generate swagger files for version 0.27.0
I generated new swagger files for the updated version, and then made all the
changes necessary to make everything compile, and make tests pass.

Most of the changes were due to swagger models storing pointers to required
fields instead of just their values.

This means we have to do a lot of pointer shenanigans and check for nils
in a bunch of places where we weren't before.

We also have some changes due to swagger's variable name generation changing.
"air-force" used to be transformed to "AirForce", but now it is "AirDashForce".

pre-commit hooks wouldn't let me split this up so it is huge.