Add some troubleshooting around `chamber env`;
I ran into this because I used `go get` to install Chamber-CLI and it
failed to work while I was using `direnv`. Users shouldn't run into this
issue anymore, but it's still possible for folks who don't install
things they way we suggest or have an old version of Chamber that
doesn't support `env` as a command option.