Fix readOnly fields breaking NTS flow
We were sending a few readOnly fields on the MTOAgent when PATCHing
MTOShipments. The new version of go-swagger rejects readOnly fields
in PATCH requests. To fix this, I removed some readOnly fields when
formatting the agent payload, and I also changed the mto agent id
to not be readOnly.
When we update info on the agent, we need the ID in order to update
the correct agent, instead of creating a new agent.

There might be a more elegant want to solve this.